Fix broken spec testing correct calling of http_status_exception and a ActionController request.
When calling a method on an instance of ActionController::Base using the #process_action method, super is called a bunch of times to call each included module in to ActionController::Base's process_action method.  A few of them - namely the Instrumentation module - look for a request object in the instance of the class, which we stuv with an empty ActionDispatch::TestRequest instance.